perf : moving sphere

+++ hittable.h
@@ -29,7 +29,7 @@ struct hit_record {
 
 class hittable {
 public:
-    virtual bool hit(const ray &r, float t_min, float t_max, hit_record &rec) const = 0;
+    virtual bool hit(const ray &r, double t_min, double t_max, hit_record &rec) const = 0;
 };
 
 #endif //RAYTRACINGTOY_HITTABLE_H
