+++ hittablelist.h
@@ -19,13 +19,13 @@ public:
 
     void add(std::shared_ptr<hittable> object) { objects.push_back(object); }
 
-    virtual bool hit(const ray &r, float t_min, float t_max, hit_record &rec) const;
+    virtual bool hit(const ray &r, double t_min, double t_max, hit_record &rec) const;
 
 public:
     std::vector<std::shared_ptr<hittable>> objects;
 };
 
-bool hittable_list::hit(const ray &r, float t_min, float t_max, hit_record &rec) const {
+bool hittable_list::hit(const ray &r, double t_min, double t_max, hit_record &rec) const {
     hit_record temp_rec;
     bool hit_anything = false;
     double closet_so_far = t_max;
